Move files into proper directory structure
