+++ src/components/Editor/Stage/Energy/Slider.jsx
@@ -1,11 +1,11 @@
 import React, {Component, PropTypes} from 'react';
 import ReactSlider from 'react-slider';
-import sliderStyles from '../../../../styles/Slider.css';
+import sliderStyles from './slider.css';
 
 export class Slider extends Component {
 
   onSlide(value){
-    this.props.setStrength(value);
+    this.props.onChange(value);
   }
 
   render() {
@@ -20,6 +20,7 @@ export class Slider extends Component {
         handleClassName={this.props.isPresentation ? " handle handle-dark" : "handle handle-light"}
         barClassName={this.props.isPresentation ? " bar bar-dark" : "bar bar-light"}
         onChange={this.onSlide.bind(this)}
+        defaultValue={this.props.value}
         value={this.props.value} />
     );
   }
@@ -29,11 +30,11 @@ export class Slider extends Component {
 Slider.propTypes = {
   isPresentation: PropTypes.bool,
   value: PropTypes.number,
-  setStrength: PropTypes.func,
+  onChange: PropTypes.func,
 };
 
 Slider.defaultProps = {
   isPresentation: false,
   value: 0,
-  setStrength(strength){},
+  onChange(strength){},
 };
