+++ src/components/Editor/Stage/Energy/slider.css
@@ -9,7 +9,7 @@
 }
 
 :global(.handle.active) {
-    background-color: lightgrey;
+  background-color: lightgrey;
 }
 
 :global(.handle-light) {
@@ -23,26 +23,26 @@
 }
 
 :global(.slider) {
-    height: 230px;
-    width: 35px;
-    border-top-left-radius: 3px;
-    border-top-right-radius: 3px;
-    box-shadow: 0px 0px 6px -2px rgba(128,128,128,1);
+  height: 230px;
+  width: 35px;
+  border-top-left-radius: 3px;
+  border-top-right-radius: 3px;
+  box-shadow: 0px 0px 6px -2px rgba(128,128,128,1);
 }
 :global(.slider-light) {
-    background-color: #fff;
+  background-color: #fff;
 }
 :global(.slider-dark) {
-    background-color: #605F5F;
+  background-color: #605F5F;
 }
 
 :global(.bar) {
-    position: relative;
-    width: 4px;
-    left: 16px;
-    top: 15px !important;
-    bottom: 15px !important;
-    border-radius: 90px !important;
+  position: relative;
+  width: 4px;
+  left: 16px;
+  top: 15px !important;
+  bottom: 15px !important;
+  border-radius: 90px !important;
 }
 :global(.bar-dark) {
   background-color: #525151;
