+++ src/components/Editor/Stage/Renderer.jsx
