+++ src/state/components/ConnectedToolbar.js
@@ -1,6 +1,6 @@
 import {connect} from 'react-redux';
 import * as actionCreators from '../action_creators';
-import {Toolbar} from '../../components/Editor/Editor/Toolbar';
+import {Toolbar} from '../../components/Editor/Toolbar';
 
 function mapStateToProps(state) {
   return {
